SampleVisitor is extracted from GUIObjectBuilder to separate class.
